requirements.txt
flask
tensorflow
opencv-python
pillow
numpy
pandas
scikit-learn
